refactor: add operation and transaction api
