+++ crates/engine/src/api/mod.rs
@@ -1,4 +1,8 @@
 /// engine trait definition
 pub(crate) mod engine_api;
+/// Storage operations trait definition;
+pub(crate) mod operation;
 /// Snapshot trait definition
 pub(crate) mod snapshot_api;
+/// Transaction trait definition;
+pub(crate) mod transaction_api;
